requirements.txt
mcp==1.9.4
uvicorn==0.34.3
python-dotenv==1.1.1
SQLAlchemy==2.0.41